move this into pg-only land
